requirements.txt
agno>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0 
sqlalchemy>=1.4.0